backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
python-dotenv
sqlparse==0.4.4
tzdata==2023.4
asgiref==3.7.2
Pillow==10.4.0
langchain==0.3.19
langchain-openai==0.3.4
langchain-community==0.3.17
langgraph==0.3.2
chromadb==0.6.3
dashscope==1.20.14
openai==1.60.1
yfinance==0.2.35
streamlit==1.32.0
zhipuai==2.1.5.20250106
duckduckgo_search>=7.5.0
pydantic>=2.10.6 
exa_py==1.8.9 
